fix: working game compilation

+++ scripts/UI/LoadingSprite.cs
@@ -4,7 +4,7 @@ using System;
 public partial class LoadingSprite : Sprite2D
 {
 
-	public override void _Process(float delta)
+	public override void _Process(double delta)
 	{
 		Rotate(0.1f);
 	}
